Backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
PyPDF2==3.0.1
docx2txt==0.8
openai==0.28.0
pymupdf==1.24.9    # optional but recommended for more robust PDF text
python-multipart==0.0.9
